+++ ai-service/src/main/java/com/example/cloudfour/aiservice/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.example.cloudfour.modulecommon.config;
+package com.example.cloudfour.aiservice.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
